Add url handling into manifest
