+++ crates/rover-client/src/operations/subgraph/publish_manifest/mod.rs
@@ -2,4 +2,4 @@ mod runner;
 mod types;
 
 pub use runner::run;
-pub use types::{SubgraphsPublishInput, SubgraphsPublishResponse, SubgraphManifest};
+pub use types::{SubgraphManifest, SubgraphsPublishInput, SubgraphsPublishResponse};
