+++ crates/rover-client/src/shared/mod.rs
