+++ src/command/subgraph/mod.rs
@@ -78,7 +78,7 @@ impl Subgraph {
             Command::Lint(command) => command.run(client_config).await,
             Command::List(command) => command.run(client_config).await,
             Command::Publish(command) => command.run(client_config, git_context).await,
-            Command::PublishManifest(command) => command.run(client_config, git_context).await
+            Command::PublishManifest(command) => command.run(client_config, git_context).await,
         }
     }
 }
