+++ src/command/subgraph/publish_manifest.rs
@@ -1,13 +1,16 @@
-use anyhow::Context;
-use clap::Parser;
-use serde::{Serialize};
-use rover_client::operations::subgraph::publish_manifest::{self, SubgraphsPublishInput, SubgraphManifest};
-use rover_client::shared::GitContext;
-use rover_std::Style;
-use crate::{RoverOutput, RoverResult};
+use crate::command::subgraph::publish_shared::{determine_routing_url, fetch_routing_url};
 use crate::options::{GraphRefOpt, ProfileOpt};
 use crate::utils::client::StudioClientConfig;
 use crate::utils::parsers::FileDescriptorType;
+use crate::{RoverOutput, RoverResult};
+use anyhow::Context;
+use clap::Parser;
+use rover_client::operations::subgraph::publish_manifest::{
+    self, SubgraphManifest, SubgraphsPublishInput,
+};
+use rover_client::shared::GitContext;
+use rover_std::Style;
+use serde::Serialize;
 
 #[derive(Debug, Serialize, Parser)]
 pub struct PublishManifest {
@@ -19,7 +22,7 @@ pub struct PublishManifest {
 
     #[serde(skip_serializing)]
     #[arg(long)]
-    manifest: FileDescriptorType
+    manifest: FileDescriptorType,
 }
 
 impl PublishManifest {
@@ -34,16 +37,25 @@ impl PublishManifest {
             .manifest
             .read_file_descriptor("subgraphs list", &mut std::io::stdin())?;
 
-        let invalid_json_err = |json| {
-            format!("JSON in {json} did not match expected format")
-        };
+        let invalid_json_err = |json| format!("JSON in {json} did not match expected format");
 
-        let subgraph_manifest =
-            serde_json::from_str::<SubgraphManifest>(&raw_manifest)
-                .with_context(|| invalid_json_err(&self.manifest))?;
+        let mut subgraph_manifest = serde_json::from_str::<SubgraphManifest>(&raw_manifest)
+            .with_context(|| invalid_json_err(&self.manifest))?;
+
+        // Determine routing urls for all graphs in subgraph_manifest
+        subgraph_manifest
+            .subgraph_inputs
+            .iter_mut()
+            .for_each(|subgraph| {
+                subgraph.url = determine_routing_url(
+                    subgraph.no_url,
+                    &subgraph.url,
+                    subgraph.allow_invalid_routing_url,
+                    fetch_routing_url(&self.graph.graph_ref, &subgraph.subgraph, &client),
+                )
+            });
 
         let subgraph_names = subgraph_manifest.get_subgraph_names();
-
         eprintln!(
             "Publishing SDL to {} (subgraphs: {}) using credentials from the {} profile.",
             Style::Link.paint(self.graph.graph_ref.to_string()),
@@ -57,8 +69,9 @@ impl PublishManifest {
                 subgraph_manifest,
                 git_context,
             },
-            &client
-        ).await?;
+            &client,
+        )
+        .await?;
 
         Ok(RoverOutput::PublishManifestResponse {
             graph_ref: self.graph.graph_ref.clone(),
@@ -66,14 +79,11 @@ impl PublishManifest {
             publish_response,
         })
 
-
-
         // eprintln!(
         //     "Publishing SDL to {} (subgraph: {}) using credentials from the {} profile.",
         //     Style::Link.paint(self.graph.graph_ref.to_string()),
         //     Style::Link.paint(&self.subgraphs.subgraph_name),
         //     Style::Command.paint(&self.profile.profile_name)
         // );
-
     }
 }
